# cmd/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "KubePi Restful API",
        "termsOfService": "https://github.com/1Panel-dev/KubePi",
        "contact": {
            "name": "Fit2cloud Support",
            "url": "https://www.fit2cloud.com",
            "email": "support@fit2cloud.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "basePath": "/kubepi/api/v1",
    "paths": {
...
        "/clusters/{name}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get cluster by name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clusters"
                ],
                "summary": "Get cluster by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "集群名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_1Panel-dev_KubePi_internal_model_v1_cluster.Cluster"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update cluster by name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clusters"
                ],
                "summary": "Update cluster by name",
                "parameters": [
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_v1_cluster.UpdateCluster"
                        }
                    },
                    {
                        "type": "string",
                        "description": "集群名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_v1_cluster.UpdateCluster"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete cluster by name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clusters"
                ],
                "summary": "Delete cluster by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "集群名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_1Panel-dev_KubePi_internal_model_v1_cluster.Cluster"
                        }
                    }
                }
            }
        },
...
        "/roles": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "List all roles",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roles"
                ],
                "summary": "List all roles",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_1Panel-dev_KubePi_internal_model_v1_role.Role"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create role",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roles"
                ],
                "summary": "Create role",
                "parameters": [
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_1Panel-dev_KubePi_internal_model_v1_role.Role"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_1Panel-dev_KubePi_internal_model_v1_role.Role"
                        }
                    }
                }
            }
        },
        "/roles/{name}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get role by name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roles"
                ],
                "summary": "Get role by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "权限名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_1Panel-dev_KubePi_internal_model_v1_role.Role"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update role by name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roles"
                ],
                "summary": "Update role by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "角色名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_1Panel-dev_KubePi_internal_model_v1_role.Role"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete role by name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roles"
                ],
                "summary": "Delete role by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "角色名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_1Panel-dev_KubePi_internal_model_v1_role.Role"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_1Panel-dev_KubePi_internal_model_v1_cluster.Authentication": {
            "type": "object",
            "properties": {
                "bearerToken": {
                    "type": "string"
                },
                "certificate": {
                    "$ref": "#/definitions/github_com_1Panel-dev_KubePi_internal_model_v1_cluster.Certificate"
                },
                "configFileContent": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "mode": {
                    "type": "string"
                }
            }
        },
        "github_com_1Panel-dev_KubePi_internal_model_v1_cluster.Certificate": {
            "type": "object",
            "properties": {
                "certData": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "keyData": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "github_com_1Panel-dev_KubePi_internal_model_v1_cluster.Cluster": {
            "type": "object",
            "properties": {
                "apiVersion": {
                    "type": "string"
                },
                "builtIn": {
                    "type": "boolean"
                },
                "caCertificate": {
                    "$ref": "#/definitions/github_com_1Panel-dev_KubePi_internal_model_v1_cluster.Certificate"
                },
                "createAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "labels": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "privateKey": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "spec": {
                    "$ref": "#/definitions/github_com_1Panel-dev_KubePi_internal_model_v1_cluster.Spec"
                },
                "status": {
                    "$ref": "#/definitions/github_com_1Panel-dev_KubePi_internal_model_v1_cluster.Status"
                },
                "updateAt": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "github_com_1Panel-dev_KubePi_internal_model_v1_cluster.Connect": {
            "type": "object",
            "properties": {
                "direction": {
                    "type": "string"
                },
                "forward": {
                    "$ref": "#/definitions/github_com_1Panel-dev_KubePi_internal_model_v1_cluster.Forward"
                }
            }
        },
        "github_com_1Panel-dev_KubePi_internal_model_v1_cluster.Forward": {
            "type": "object",
            "properties": {
                "apiServer": {
                    "type": "string"
                },
                "proxy": {
                    "$ref": "#/definitions/github_com_1Panel-dev_KubePi_internal_model_v1_cluster.Proxy"
                }
            }
        },
        "github_com_1Panel-dev_KubePi_internal_model_v1_cluster.Proxy": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "github_com_1Panel-dev_KubePi_internal_model_v1_cluster.Spec": {
            "type": "object",
            "properties": {
                "authentication": {
                    "$ref": "#/definitions/github_com_1Panel-dev_KubePi_internal_model_v1_cluster.Authentication"
                },
                "connect": {
                    "$ref": "#/definitions/github_com_1Panel-dev_KubePi_internal_model_v1_cluster.Connect"
                },
                "local": {
                    "type": "boolean"
                }
            }
        },
        "github_com_1Panel-dev_KubePi_internal_model_v1_cluster.Status": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "phase": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
...
        "github_com_1Panel-dev_KubePi_internal_model_v1_role.PolicyRule": {
            "type": "object",
            "properties": {
                "resource": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "resourceNames": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "verbs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "github_com_1Panel-dev_KubePi_internal_model_v1_role.Role": {
            "type": "object",
            "properties": {
                "apiVersion": {
                    "type": "string"
                },
                "builtIn": {
                    "type": "boolean"
                },
                "createAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "rules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_1Panel-dev_KubePi_internal_model_v1_role.PolicyRule"
                    }
                },
                "updateAt": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
...
        "internal_api_v1_cluster.UpdateCluster": {
            "type": "object",
            "properties": {
                "apiServer": {
                    "type": "string"
                },
                "certData": {
                    "type": "string"
                },
                "configFileContent": {
                    "type": "string"
                },
                "keyData": {
                    "type": "string"
                },
                "labels": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "mode": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "withLabel": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}